edge-ai-void-detection/edgedevice/modules/imageprocessor/requirements.txt
azure-iothub-device-client~=1.4.3
azure.storage.blob~=2.1.0
google
tensorflow~=1.14.0
tensorflow-serving-api
grpcio
grpcio-tools
numpy
